[Billboard] Creating a billboard effect, used to display text in a 3d scene (implemented on the basic axis object) + many little improvments of classes like Camera and Object

+++ Nc/Graphics/Object/String.h
@@ -62,7 +62,7 @@ namespace Nc
                 typedef std::map<std::string, Font*>    MapFont;
 
             public:
-                String(const Utils::Unicode::UTF32 &text, unsigned int size, const Color &color, const std::string &ttf, const Utils::Mask<Style> &s = Regular);
+                String(const Utils::Unicode::UTF32 &text, float size, const Color &color, const std::string &ttf, const Utils::Mask<Style> &s = Regular);
                 virtual ~String();
 
                 static const char               *ClassName()                                    {return "String";}
